moved all tile tipe classes in to TileType
